requirements.txt
# Core utilities
python-dotenv==1.0.1
pandas==2.3.3
numpy==2.3.4

# PDF parsing (text only, no OCR)
PyMuPDF==1.26.5          

# Vector store 
chromadb==1.2.1      

# OpenAI API (embeddings + chat)
openai==2.3.0           

# Tokenization / chunking
tiktoken==0.7.0

# Frontend UI
streamlit==1.39.0

# Progress
tqdm==4.67.1
